Articles Summary Automation
I have created an automation for the news feeds such as US news, Travel news, Science news, DIY news, Movies news, Podcasts news. in this automation whenever the news is uploaded to the page, rss triggers the initial call for process to begin. and hence it checks if the article is already summarized in the Google Sheet. if not, appends it into the sheet. before that the summarization includes the below format to store the news.
{
  "Title": "",
  "Context/Problem": "",
  "The Innovation": "",
  "Key Features": [],
  "Applications": "",
  "Challenges": "",
  "Future Outlook": "",
  "Conclusion": ""
}
 in this format it gives you the complete insights and as well summarized version of the news in under single column.